requirements.txt
Django
djangorestframework
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.2.0
django-storages==1.14.2
django-filter
requests
psycopg2-binary
reportlab
openpyxl
gunicorn
dotenv
boto3
botocore
drf-spectacular
drf-spectacular-sidecar
python-decouple==3.8
pinecone
openai
tiktoken
stripe
pandas